slice not splice
If I understand what you were trying to do,
was to get 107 shelters to pass
into the other component.

Array.splice will modify the initial array,
and at times that isn't the behaviour
that you wanted.

Slice, will return a new array - with the
subset of items that you want, so
array.slice(0,107) will get the those,